Convert in order linker into proxy component.
