+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/wiring/PlatformCoordinator.java
@@ -22,11 +22,13 @@ import com.swirlds.platform.components.consensus.ConsensusEngine;
 import com.swirlds.platform.event.GossipEvent;
 import com.swirlds.platform.event.creation.EventCreationManager;
 import com.swirlds.platform.event.deduplication.EventDeduplicator;
+import com.swirlds.platform.event.linking.InOrderLinker;
 import com.swirlds.platform.event.orphan.OrphanBuffer;
 import com.swirlds.platform.event.validation.EventSignatureValidator;
 import com.swirlds.platform.event.validation.InternalEventValidator;
 import com.swirlds.platform.eventhandling.TransactionPrehandler;
 import com.swirlds.platform.internal.ConsensusRound;
+import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.system.events.BaseEventHashedData;
 import com.swirlds.platform.wiring.components.ConsensusRoundHandlerWiring;
 import com.swirlds.platform.wiring.components.ShadowgraphWiring;
@@ -52,7 +54,7 @@ public class PlatformCoordinator {
     private final ComponentWiring<EventDeduplicator, GossipEvent> eventDeduplicatorWiring;
     private final ComponentWiring<EventSignatureValidator, GossipEvent> eventSignatureValidatorWiring;
     private final ComponentWiring<OrphanBuffer, List<GossipEvent>> orphanBufferWiring;
-    private final InOrderLinkerWiring inOrderLinkerWiring;
+    private final ComponentWiring<InOrderLinker, EventImpl> inOrderLinkerWiring;
     private final ShadowgraphWiring shadowgraphWiring;
     private final ComponentWiring<ConsensusEngine, List<ConsensusRound>> consensusEngineWiring;
     private final ComponentWiring<EventCreationManager, BaseEventHashedData> eventCreationManagerWiring;
@@ -84,7 +86,7 @@ public class PlatformCoordinator {
             @NonNull final ComponentWiring<EventDeduplicator, GossipEvent> eventDeduplicatorWiring,
             @NonNull final ComponentWiring<EventSignatureValidator, GossipEvent> eventSignatureValidatorWiring,
             @NonNull final ComponentWiring<OrphanBuffer, List<GossipEvent>> orphanBufferWiring,
-            @NonNull final InOrderLinkerWiring inOrderLinkerWiring,
+            @NonNull final ComponentWiring<InOrderLinker, EventImpl> inOrderLinkerWiring,
             @NonNull final ShadowgraphWiring shadowgraphWiring,
             @NonNull final ComponentWiring<ConsensusEngine, List<ConsensusRound>> consensusEngineWiring,
             @NonNull final ComponentWiring<EventCreationManager, BaseEventHashedData> eventCreationManagerWiring,
@@ -128,7 +130,7 @@ public class PlatformCoordinator {
         eventDeduplicatorWiring.flush();
         eventSignatureValidatorWiring.flush();
         orphanBufferWiring.flush();
-        inOrderLinkerWiring.flushRunnable().run();
+        inOrderLinkerWiring.flush();
         shadowgraphWiring.flushRunnable().run();
         consensusEngineWiring.flush();
         applicationTransactionPrehandlerWiring.flush();
@@ -176,7 +178,7 @@ public class PlatformCoordinator {
         // Data is no longer moving through the system. Clear all the internal data structures in the wiring objects.
         eventDeduplicatorWiring.getInputWire(EventDeduplicator::clear).inject(NoInput.getInstance());
         orphanBufferWiring.getInputWire(OrphanBuffer::clear).inject(NoInput.getInstance());
-        inOrderLinkerWiring.clearInput().inject(NoInput.getInstance());
+        inOrderLinkerWiring.getInputWire(InOrderLinker::clear).inject(NoInput.getInstance());
         stateSignatureCollectorWiring.getClearInput().inject(NoInput.getInstance());
         eventCreationManagerWiring.getInputWire(EventCreationManager::clear).inject(NoInput.getInstance());
     }
